requirements.txt
cuda12.1
cudnn==8.9.2
python==3.11.8
pytorch==2.2.2
dgl-cuda12.1==2.2.0
transformers==4.39.3
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.0
yaml==6.0.1